a) update object mapping i) move Message, Sms and Status to common package ii) remove ConfigRecordMessageConverter class iii) use default spring object mapping for serialization b) minor updates and typos on documentation
