+++ src/main/resources/application.yml
@@ -15,7 +15,9 @@ spring:
       bootstrap-servers: ${KAFKA_SERVERS:localhost:9092}
       group-id: group_id
       auto-offset-reset: earliest
-      value-deserializer: org.apache.kafka.common.serialization.ByteArrayDeserializer
+      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
+      properties:
+        spring.json.type.mapping: message:com.pilot.commons.Message, sms:com.pilot.commons.Sms
 
 ---
 #'producer' profile only
@@ -31,7 +33,7 @@ spring:
       bootstrap-servers: ${KAFKA_SERVERS:localhost:9092}
       value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
       properties:
-        spring.json.type.mapping: message:com.pilot.springbootkafkaproducerconsumer.model.Message, sms:com.pilot.springbootkafkaproducerconsumer.model.Sms
+        spring.json.type.mapping: message:com.pilot.commons.Message, sms:com.pilot.commons.Sms
 
 server:
   port: 9000
